refactor: restructure knowledge stack for clarity
Extract note link graph logic into links.py, add managed_connection()
context manager, and deduplicate repeated patterns.

Changes:
- Extract links.py from ingest.py (wikilink resolution + similarity graph)
- Add managed_connection() context manager to database.py, eliminating
  the own_conn/try/finally boilerplate from 4 call sites
- Deduplicate validate_score logic in models.py via shared helper
- Deduplicate _git() closure in save.py into module-level function
- Remove duplicate _require_document_id from ingest.py (use database.py's)
- Split test_links.py from test_ingest.py to match new module structure

Before/after (key files):
  ingest.py:    555 → 352 lines (-37%)
  links.py:    new, 200 lines
  test_ingest:  916 → 739 lines
  test_links:  new, 181 lines

All 91 tests pass, zero regressions. Public API unchanged.
