+++ stacks/knowledge/app/knowledge/__init__.py
@@ -13,6 +13,7 @@ from .database import (
     list_documents,
     list_documents_by_source_prefix,
     list_related_documents,
+    managed_connection,
     resolve_database_url,
     run_migrations,
     search_chunks,
@@ -61,6 +62,7 @@ __all__ = [
     "list_documents",
     "list_documents_by_source_prefix",
     "list_related_documents",
+    "managed_connection",
     "related",
     "resolve_database_url",
     "run_migrations",
