+++ stacks/knowledge/app/tests/test_ingest.py
@@ -9,13 +9,9 @@ import pytest
 
 from knowledge import __main__ as cli
 from knowledge.ingest import (
-    _all_similarity_note_links,
     _extract_pdf_text,
     _file_content_hash,
     _read_file_content,
-    _refresh_similarity_note_links,
-    _resolved_wikilink_targets,
-    _similarity_note_links,
     _title_from_file,
     ingest_directory,
     ingest_file,
@@ -26,8 +22,6 @@ from knowledge.models import (
     DirectoryIngestResult,
     Document,
     IngestResult,
-    NoteLink,
-    RelatedDocument,
 )
 
 
@@ -47,10 +41,10 @@ def _task_event(stderr: str) -> dict[str, object]:
     return json.loads(stderr.strip().splitlines()[-1])
 
 
-@patch("knowledge.ingest.connect")
+@patch("knowledge.database.connect")
 @patch("knowledge.ingest.get_embeddings", side_effect=_fake_embeddings)
 @patch("knowledge.ingest.insert_chunks", return_value=[])
-@patch("knowledge.ingest._refresh_note_links")
+@patch("knowledge.ingest.refresh_note_links")
 @patch("knowledge.ingest.upsert_document")
 @patch("knowledge.ingest.delete_document_chunks")
 @patch("knowledge.ingest.get_document_by_source", return_value=None)
@@ -89,10 +83,10 @@ def test_ingest_file_new_document(
     mock_refresh_note_links.assert_called_once()
 
 
-@patch("knowledge.ingest.connect")
+@patch("knowledge.database.connect")
 @patch("knowledge.ingest.get_embeddings", side_effect=_fake_embeddings)
 @patch("knowledge.ingest.insert_chunks", return_value=[])
-@patch("knowledge.ingest._refresh_note_links")
+@patch("knowledge.ingest.refresh_note_links")
 @patch("knowledge.ingest.upsert_document")
 @patch("knowledge.ingest.delete_document_chunks")
 @patch("knowledge.ingest.get_document_by_source")
@@ -133,10 +127,10 @@ def test_ingest_skips_unchanged_content(
     mock_refresh_note_links.assert_not_called()
 
 
-@patch("knowledge.ingest.connect")
+@patch("knowledge.database.connect")
 @patch("knowledge.ingest.get_embeddings", side_effect=_fake_embeddings)
 @patch("knowledge.ingest.insert_chunks", return_value=[])
-@patch("knowledge.ingest._refresh_note_links")
+@patch("knowledge.ingest.refresh_note_links")
 @patch("knowledge.ingest.upsert_document")
 @patch("knowledge.ingest.delete_document_chunks")
 @patch("knowledge.ingest.get_document_by_source")
@@ -174,10 +168,10 @@ def test_ingest_reingests_changed_content(
     mock_refresh_note_links.assert_called_once()
 
 
-@patch("knowledge.ingest.connect")
+@patch("knowledge.database.connect")
 @patch("knowledge.ingest.get_embeddings", side_effect=_fake_embeddings)
 @patch("knowledge.ingest.insert_chunks", return_value=[])
-@patch("knowledge.ingest._refresh_note_links")
+@patch("knowledge.ingest.refresh_note_links")
 @patch("knowledge.ingest.upsert_document")
 @patch("knowledge.ingest.delete_document_chunks")
 @patch("knowledge.ingest.get_document_by_source", return_value=None)
@@ -210,10 +204,10 @@ def test_ingest_text_uses_text_uri(
     mock_refresh_note_links.assert_called_once()
 
 
-@patch("knowledge.ingest.connect")
+@patch("knowledge.database.connect")
 @patch("knowledge.ingest.get_embeddings", side_effect=_fake_embeddings)
 @patch("knowledge.ingest.insert_chunks", return_value=[])
-@patch("knowledge.ingest._refresh_note_links")
+@patch("knowledge.ingest.refresh_note_links")
 @patch("knowledge.ingest.upsert_document")
 @patch("knowledge.ingest.delete_document_chunks")
 @patch("knowledge.ingest.get_document_by_source", return_value=None)
@@ -251,10 +245,10 @@ def test_ingest_text_different_content_same_title_no_collision(
     assert mock_refresh_note_links.call_count == 2
 
 
-@patch("knowledge.ingest.connect")
+@patch("knowledge.database.connect")
 @patch("knowledge.ingest.get_embeddings", side_effect=_fake_embeddings)
 @patch("knowledge.ingest.insert_chunks", return_value=[])
-@patch("knowledge.ingest._refresh_note_links")
+@patch("knowledge.ingest.refresh_note_links")
 @patch("knowledge.ingest.upsert_document")
 @patch("knowledge.ingest.delete_document_chunks")
 @patch("knowledge.ingest.get_document_by_source")
@@ -294,178 +288,7 @@ def test_zero_chunks_on_reingest_deletes_stale_data(
     mock_refresh_note_links.assert_called_once()
 
 
-def test_resolved_wikilink_targets_match_case_insensitive_partial_paths() -> None:
-    # Arrange
-    source = Document(
-        id=UUID("00000000-0000-0000-0000-000000000001"),
-        source_path="/notes/source.md",
-        title="Source",
-        content_hash="hash-source",
-    )
-    first_target = Document(
-        id=UUID("00000000-0000-0000-0000-000000000002"),
-        source_path="/notes/projects/Alpha Note.md",
-        title="Alpha",
-        content_hash="hash-alpha",
-    )
-    second_target = Document(
-        id=UUID("00000000-0000-0000-0000-000000000003"),
-        source_path="/notes/reference/beta.md",
-        title="Beta",
-        content_hash="hash-beta",
-    )
-    content = (
-        "See [[projects/alpha note]], [[BETA]], [[Alpha Note#Overview|alpha]], and [[Missing]]."
-    )
-
-    # Act
-    targets = _resolved_wikilink_targets(
-        content,
-        source_document=source,
-        documents=[source, first_target, second_target],
-    )
-
-    # Assert
-    assert [document.source_path for document in targets] == [
-        "/notes/projects/Alpha Note.md",
-        "/notes/reference/beta.md",
-        "/notes/projects/Alpha Note.md",
-    ]
-
-
-@patch("knowledge.ingest.find_similar_documents")
-def test_similarity_note_links_add_bidirectional_edges(
-    mock_find_similar_documents: MagicMock,
-) -> None:
-    # Arrange
-    source_id = UUID("00000000-0000-0000-0000-000000000001")
-    target_id = UUID("00000000-0000-0000-0000-000000000002")
-    source = Document(
-        id=source_id,
-        source_path="docs/source.md",
-        title="Source",
-        content_hash="hash-source",
-    )
-    target = Document(
-        id=target_id,
-        source_path="docs/target.md",
-        title="Target",
-        content_hash="hash-target",
-    )
-    mock_find_similar_documents.return_value = [
-        RelatedDocument(
-            link_type="similarity",
-            score=0.91,
-            document=target,
-        )
-    ]
-
-    # Act
-    links = _similarity_note_links(MagicMock(), source_document=source)
-
-    # Assert
-    assert {(link.source_id, link.target_id, link.score) for link in links} == {
-        (source.id, target.id, 0.91),
-        (target.id, source.id, 0.91),
-    }
-
-
-@patch("knowledge.ingest._similarity_note_links")
-@patch("knowledge.ingest.list_documents")
-def test_all_similarity_note_links_deduplicates_edges(
-    mock_list_documents: MagicMock,
-    mock_similarity_note_links: MagicMock,
-) -> None:
-    # Arrange
-    source_id = UUID("00000000-0000-0000-0000-000000000001")
-    target_id = UUID("00000000-0000-0000-0000-000000000002")
-    source = Document(
-        id=source_id,
-        source_path="docs/source.md",
-        title="Source",
-        content_hash="hash-source",
-    )
-    target = Document(
-        id=target_id,
-        source_path="docs/target.md",
-        title="Target",
-        content_hash="hash-target",
-    )
-    mock_list_documents.return_value = [source, target]
-    mock_similarity_note_links.side_effect = [
-        [
-            NoteLink(
-                source_id=source_id,
-                target_id=target_id,
-                link_type="similarity",
-                score=0.91,
-            ),
-            NoteLink(
-                source_id=target_id,
-                target_id=source_id,
-                link_type="similarity",
-                score=0.91,
-            ),
-        ],
-        [
-            NoteLink(
-                source_id=target_id,
-                target_id=source_id,
-                link_type="similarity",
-                score=0.91,
-            ),
-            NoteLink(
-                source_id=source_id,
-                target_id=target_id,
-                link_type="similarity",
-                score=0.91,
-            ),
-        ],
-    ]
-    conn = MagicMock()
-
-    # Act
-    links = _all_similarity_note_links(conn)
-
-    # Assert
-    assert {(link.source_id, link.target_id) for link in links} == {
-        (source_id, target_id),
-        (target_id, source_id),
-    }
-    mock_list_documents.assert_called_once_with(conn)
-    assert [call.args[0] for call in mock_similarity_note_links.call_args_list] == [conn, conn]
-    source_documents = [
-        call.kwargs["source_document"] for call in mock_similarity_note_links.call_args_list
-    ]
-    assert source_documents == [source, target]
-
-
-@patch("knowledge.ingest.insert_note_links")
-@patch("knowledge.ingest.delete_note_links")
-@patch("knowledge.ingest._all_similarity_note_links")
-def test_refresh_similarity_note_links_rebuilds_graph(
-    mock_all_similarity_note_links: MagicMock,
-    mock_delete_note_links: MagicMock,
-    mock_insert_note_links: MagicMock,
-) -> None:
-    # Arrange
-    conn = MagicMock()
-    links = [
-        MagicMock(),
-        MagicMock(),
-    ]
-    mock_all_similarity_note_links.return_value = links
-
-    # Act
-    _refresh_similarity_note_links(conn)
-
-    # Assert
-    mock_delete_note_links.assert_called_once_with(conn, link_type="similarity")
-    mock_all_similarity_note_links.assert_called_once_with(conn)
-    mock_insert_note_links.assert_called_once_with(conn, links)
-
-
-@patch("knowledge.ingest.connect")
+@patch("knowledge.database.connect")
 @patch("knowledge.ingest.delete_document", return_value=1)
 @patch("knowledge.ingest.list_documents_by_source_prefix")
 @patch("knowledge.ingest.ingest_file")
@@ -524,7 +347,7 @@ def test_ingest_directory_reuses_connection_and_deletes_orphans(
     conn.close.assert_called_once()
 
 
-@patch("knowledge.ingest.connect")
+@patch("knowledge.database.connect")
 @patch("knowledge.ingest.delete_document", return_value=0)
 @patch("knowledge.ingest.list_documents_by_source_prefix", return_value=[])
 @patch("knowledge.ingest.ingest_file")
@@ -570,7 +393,7 @@ def test_ingest_directory_continues_after_file_error(
     mock_list_prefix.assert_called_once()
 
 
-@patch("knowledge.ingest.connect")
+@patch("knowledge.database.connect")
 @patch("knowledge.ingest.delete_document", return_value=1)
 @patch("knowledge.ingest.list_documents_by_source_prefix")
 @patch("knowledge.ingest.ingest_file")
@@ -829,10 +652,10 @@ def test_read_file_content_markdown(tmp_path: Path) -> None:
     assert result == "# Hello\n\nWorld"
 
 
-@patch("knowledge.ingest.connect")
+@patch("knowledge.database.connect")
 @patch("knowledge.ingest.get_embeddings", side_effect=_fake_embeddings)
 @patch("knowledge.ingest.insert_chunks", return_value=[])
-@patch("knowledge.ingest._refresh_note_links")
+@patch("knowledge.ingest.refresh_note_links")
 @patch("knowledge.ingest.upsert_document")
 @patch("knowledge.ingest.delete_document_chunks")
 @patch("knowledge.ingest.get_document_by_source", return_value=None)
@@ -892,7 +715,7 @@ def test_pdf_hash_uses_raw_bytes(tmp_path: Path) -> None:
     assert _file_content_hash(pdf_a) != _file_content_hash(pdf_b)
 
 
-@patch("knowledge.ingest.connect")
+@patch("knowledge.database.connect")
 @patch("knowledge.ingest.get_embeddings", side_effect=RuntimeError("API down"))
 @patch("knowledge.ingest.get_document_by_source", return_value=None)
 def test_ingest_file_rolls_back_shared_connection_on_error(
